Refactor of build scripts
This commit refactors some scripts run to build
the project in a docker container. Among the modifications

-  Refactor of build-install.sh
-  Manual installation of boost
-  Add a flag such that one can opt to generate coverage
files or not

Refactor travis build scripts

This commit splits the docker file used to build
the project in two, one for binaries (dockerfile.bins,
which is used to build a node.js server) and another
for testing purposes (dockerfile.travis).