[server] Add basic error handling
This commit sets up a basic interface to inform the user
about an internal error.

All fatal errors are redirected to the generic error page,
but the error message is registered in the server.

Non fatal errors simply display an alert in the browser.
At this point, this could happen if the definition of
some word is not found.