[server] Add no-puzzle-generated error
This commit adds a custom error whenever a puzzle could
not be generated due to the length of the input string.

Additionally, this commit also fixes an error in the POST
request for /api/puzzle-from-file. The setting of header
'Content-Type' was not correct.

When sending a file, `multipart/form-data` needs an extra
setting: boundary=<something>. This is generated automatically
if we don't explicit set the header. The manual setting overwrites
the good setting of `Content-Type`.