Modify regex in find_path algorithm
This commits modifies the regex used to match words
in a flat representation of a brick. Previously, several
words were being skipped because of overlapping matches.
The new regex uses lookbehind and lookahead.

Additionally, it a new parameter in the find_path algorithm
that allows enumeration of all possible paths. It is disabled
by default. A new test was added to test this functionality.